tracing: disable regular expression matching in log filters (#1580)
[`tracing-subscriber` v0.3.10][1] introduces a new [builder API][2] for
configuring the `EnvFilter` type. One of the configurations that can now
be set using the builder is whether span field value filters for
`fmt::Debug` fields are interpreted as precise string matching or as
regular expressions.

Disabling regular expressions is strongly recommended in cases where
filter configurations can come from an untrusted source, as a malicious
regular expression is a potential denial-of-service vector. In the
proxy, we already implement a form of access control for setting the
filter --- the `/admin/log-level` endpoint denies requests that did not
originate from localhost, so it's only possible to set the log level
when SSHed into the pod. However, it's probably wise to disable regex
filters here as well, as a form of additional defense in depth.

Therefore, this branch updates the `tracing-subscriber` dependency to
v0.3.10, and disables regular expression filters.

[1]: https://github.com/tokio-rs/tracing/releases/tag/tracing-subscriber-0.3.10
[2]: https://docs.rs/tracing-subscriber/latest/tracing_subscriber/filter/struct.Builder.html