Split out list-changed-crates into an action
Our check-each workflow is subtly broken: we specify a YAML list of
paths to watch for changes, but the `changed-files` action expects a
newline-separated list. This change fixes this configuration.

This change also moves the complicated `list-changed-crates` script into
its own (local) action so that the script is easier to read/test/maintain.

Signed-off-by: Oliver Gould <ver@buoyant.io>