metrics: add `process_uptime_seconds_total` metric (#1659)
The proxy exposes a `process_cpu_seconds_total` metric. It also exposes
a `process_start_time_seconds` metric. But given a random metrics
snapshot, we have no way of knowing the proxy's uptime unless we know
the current time at the time of the snapshot. linkerd/linkerd2#8443
proposed adding a `process_uptime_seconds_total` metric tracking the
(fractional) number of seconds since the process started, to make these
other process metrics easier to interpret.

This branch implements the `process_uptime_seconds_total` metric.
Actually adding the metric is rather simple.

However, it occurred to me that uptimes are also used as timestamps in
the proxy's `tracing` logs. Currently, the start time used to calculate
log timestamps is just recorded when constructing the logger. If
separate `Instant`s are used as the start time for log uptime timestamps
and for the `process_uptime_seconds_total` metric, this means the
process' uptime as reported by the metrics endpoint and the process'
uptime according to log messages may not quite agree. Therefore, I
thought it would be good if a single `Instant` was used for both logging
and metrics, and I made some additional changes so that we can record a
single timestamp when the proxy starts, and pass it into both the
metrics endpoint and the logger. This does make this branch a slightly
larger change, though.

Closes linkerd/linkerd2#8443

Signed-off-by: Eliza Weisman <eliza@buoyant.io>