dns: Fall back to A record when SRV resolution fails (#1670)
When resolving control plane components, we first resolve a SRV
record. If we're unable to parse IPv4 addresses from the record,
then we fall back to an A record lookup. But if the initial resolution
fails completely (i.e. doesn't produce any record data), then we don't
fall back to an A record.

This change fixes this fallback behavior so that we attempt A record
lookups whenever a SRV lookup fails. In doing so, we change the dns
module to use typed errors (avoiding boxing) so that we can print
better error messages and have static guarantees about the error
types being handled.

Fixes linkerd/linkerd2#8296

Signed-off-by: Oliver Gould <ver@buoyant.io>
Co-authored-by: Alex Leong <alex@buoyant.io>