cache: generalize `Cache` into a map structure
Currently, the `linkerd-cache` crate contains an implementation of a
cache where entries are evicted if they have not been accessed for a
timeout duration. This type currently implements the `NewService` trait
and requires that cached values are `Service`s. It does not provide a
more general purpose key-value cache interface.

PR #1677 introduces dynamic discovery for inbound port policies. We
would like to implement a similar eviction strategy for these policy
watches. However, the policy watches are *not* used in a `NewService`
implementation, but in a `CheckPolicy` implementation.

Therefore, this branch rewrites the `Cache` type to serve as a
general-purpose key-value cache, and adds a new `NewCachedService` type
that implements `NewService` on top of `Cache`.