Store permanent values in the cache (for #1683)
In some cases--like inbound policy discovery--we have a set of permanent
values that should never be removed from the cache. This change updates
the cache to only hold a handle when eviction is scheduled. No handle is
held when the cache entry is permanent.

This change also fixes a possible race condition. Previously, the
eviction task could consume an entry's handle without locking the cache.
This made it possible for a `get_or_insert_with` call to race against
the eviction task, adding an entry that would be immediately evicted by
the eviction task.

Signed-off-by: Oliver Gould <ver@buoyant.io>