dev: Replace `Makefile` with `justfile`
Other projects in the ecosystem (like kube-rs) are ditching Make's
obtuse syntax for a newer tool called [`just`][just].

This change removes our `Makefile` with a simpler `justfile`. The
release workflow is updated to use `just`. We also remove our local
`package` action, as it's not reused.

While doing this, a new version of the devcontainer has been published
with `just`, `cargo-nextest`, and various dependency updates.

[just]: https://just.systems/man/

Signed-off-by: Oliver Gould <ver@buoyant.io>