test: replace `profile_test!` macro with builder (#1705)
Currently, the service profiles integration test uses a very complex
declarative macro to generate test bodies. This approach has some
downsides:

1. All of the test code is contained inside a macro. This means that
   `rustfmt` will not reformat it, and (in some cases) compilation
   errors can be unclear.
2. The macro-generated code is relatively inflexible. This makes it
   difficult to add new tests that need to do slightly different things.
3. Adding new configurations to the macro is complex. If we want to add
   some new thing that can also be passed in to change the generated
   code, we have to modify a bunch of macro match arms, which is
   annoying and time-consuming.

This branch rewrites the profiles integration tests to use a
builder-style design instead of a macro. This way, the repetitive
boilerplate necessary to set up the tests is still factored out, and can
still be overridden by passing in different parameters to the builder,
but instead of having a giant macro generate the test, the test itself
is just normal Rust code.

The builder spins up the test proxy and servers, and returns the test
client and metrics client to the test itself, and the test can now do
whatever it wants with them, rather than having to pass bits of test
code into a macro. This is much nicer, because the tests are just normal
tests now. Additionally, the builder can accept an arbitrary server
configuration, which will be necessary for some new tests I'm working on
in a separate branch.

I've also changed the profiles integration test so that most of the
tests are run with both HTTP/1 *and* HTTP/2 clients and servers. This is
still accomplished using macros, but now, the macro is much simpler. The
tests themselves are factored out into functions that take either an
HTTP/1 or HTTP/2 server, and all the macro does is generate wrapper
functions that call the actual test function with the appropriate server
version. This way, the actual test code is never inside of a macro
expansion.

I'd like to change more of the integration tests to work similarly to
these, removing other cases where a macro is used to generate test
boilerplate, and replacing it with a builder.  Also, and more importantly,
this change is necessary for some additional tests I'll be adding in a 
separate PR.

Signed-off-by: Eliza Weisman <eliza@buoyant.io>