justfile: generate a default docker tag name
Currently, when running `just docker`, the image is not given a
meaningful tag name by default. This commit changes the `justfile` so
that the default tag is something useful, rather than `''`. If no tag
name is provided from the command line, the Docker image will be given
the tag `gchr.io/linkerd-io/proxy:$USER-$GIT_SHA`. The image name of the
actual Linkerd repository is used so that dev images can be selected
using only the `--proxy-version` flag of `linkerd inject`, and
`--proxy-image` is not needed.

If we'd prefer not to output the same image name as release images by
default, I could that part to something different.