justfile: generate a default docker tag name (#1716)
Currently, when running `just docker`, the image is not given a
meaningful tag name by default. This commit changes the `justfile` so
that the default tag is something useful, rather than `''`. If no tag
name is provided from the command line, the Docker image will be given
the tag `localhost/linkerd/proxy:<branch>.<sha>`.

A `features` variable is also set which can be used to enable
alternate proxy features. 

Signed-off-by: Eliza Weisman <eliza@buoyant.io>
Signed-off-by: Oliver Gould <ver@buoyant.io>
Co-authored-by: Oliver Gould <ver@buoyant.io>