stack: add `Proxy::proxy_oneshot`
Currently, `ServiceExt::oneshot` is used to take a `Service` by value,
poll it until it's ready, and then call that service. This composes well
with `Service` middleware, but it does not compose with our `Proxy`
trait. If a `Service` must be wrapped in a `Proxy` call, it is
impossible to oneshot it.

This branch adds a new `Proxy::proxy_oneshot` method to `Proxy` that
takes a `Proxy` and a `Service` by value, and returns a `Future` that
drives the `Service` to readiness and then calls it through that
`Proxy`.

Currently, this is unused, but it will be used in PR #1706 (from which
this change was factored out).