http-box: add `EraseResponse` middleware (#1723)
The `linkerd-http-box` crate currently contains `BoxRequest` and
`BoxResponse` middleware for boxing request/response bodies, and an
`EraseRequest` middleware for erasing the type of request bodies, for
use in a stack which must be able to serve *multiple* request body
types. However, there is no corresponding `EraseResponse` middleware.

This branch adds an `EraseResponse` middleware in `linkerd-http-box`.
This is necessary for PR #1706, but I factored this out into its own
branch as it's a fairly self-contained change.