test: don't generate discovery tests in a macro (#1711)
Currently, the service discovery integration tests are generated using a
giant macro that expands to all of the different tests. The macro is
used to run the tests with both HTTP/1 (with and without absolute URIs)
and HTTP/2 clients/servers. This approach has some downsides:

1. All of the test code is contained inside a macro. This means that
   `rustfmt` will not reformat it, and (in some cases) compilation
   errors can be unclear.
2. The macro-generated code is relatively inflexible. This makes it
   difficult to add new tests that need to do slightly different things.
3. Adding new configurations to the macro is complex. If we want to add
   some new thing that can also be passed in to change the generated
   code, we have to modify a bunch of macro match arms, which is
   annoying and time-consuming.

This PR changes these tests similarly to the way #1705 modified the
profiles integration tests. Rather than generating the tests themselves
inside the macro, we define version-agnostic test bodies in a
`cross_version` module, which accept the HTTP version configuration as
an argument. Now, a much simpler macro is used to just generate simple
test wrappers that call into the cross-version test definitions. This
means that none of the actual test code is inside of a macro expansion.

Depends on #1705

Signed-off-by: Eliza Weisman <eliza@buoyant.io>