Implement `Deref` for address newtypes
We have several address-related newtypes that wrap `SocketAddr` types.
They currently implement `AsRef<SocketAddr>`, but this is a little
cumbersome in practice. This change replaces these `AsRef`
implementations with `Deref` so that we can, for instance, call `.ip()`
and `.port()` on these types directly without special handling.

No functional changes.

Signed-off-by: Oliver Gould <ver@buoyant.io>