Add an HTTP route redirect filter
The Gateway API requires that `HTTPRoute` implementations support a
[redirection filter][gwapi].

This change adds a filter implementation to the `http-route` crate.
Inbound server policies are updated to apply this filter on HTTP routes.
When a request is redirected, the policy layer returns an error and the
error crate is responsible for turning that error into an HTTP response.

This change fixes path matching so that trailing slashes do not impact
prefix matching.

[gwapi]: https://github.com/kubernetes-sigs/gateway-api/blob/541e9fc2b3c2f62915cb58dc0ee5e43e4096b3e2/apis/v1beta1/httproute_types.go#L772-L821

Signed-off-by: Oliver Gould <ver@buoyant.io>