Configure inbound HTTP routes via gRPC
The policy controller serves inbound server configuration. Recent
changes have updated the inbound proxy to support applying HTTP
route-specific policies, but only a single default route configuration
was used.

This change updates the proxy to use a (not yet released) new proxy API
version that includes HTTP routes in server responses. This change adds
protobuf conversions from these types to the proxy's HTTP route types.

If the proxy receives a router filter of an unknown type (i.e., because
the controller is running a later version of the API that includes new
types), then the proxy will FAIL all requests on that route with an
internal server error. It's considered safer to fail hard in this case,
rather than to silently ignore a configured policy that could
potentially be security-sensitive.

Signed-off-by: Oliver Gould <ver@buoyant.io>