chore: update Rust version to 1.62.1
A patch release of stable Rust, v1.62.1, [was released today][1]. This
patch fixes compiler regressions in 1.62.0.

This branch updates the proxy to use 1.62.1 rather than 1.62.0. For me,
at least, this is necessary to make Cargo commands work in this repo ---
when the `rust-version` file is set to 1.62.0, running cargo commands
results in an error like

```
:; cargo test
error: the 'cargo' binary, normally provided by the 'cargo' component, is not applicable to the '1.62.0-x86_64-unknown-linux-gnu' toolchain
```

[1]: https://blog.rust-lang.org/2022/07/19/Rust-1.62.1.html