dev: Reuse the linkerd dev container
The proxy repo maintains a nearly identitical devcontainer configuration
to that used in the main linkerd2 repo (minus some Go tooling). This
means that we have to pull the same sets of tools redundantly, and that
it's easy for the tool versions to diverge needlessly.

This change updates the proxy to use the published linkerd dev tools
image as a base container, adding a rust nightly toolchain for
cargo-fuzz (which is only used in this repo currently).

Depends on linkerd/linkerd2#8925.

Signed-off-by: Oliver Gould <ver@buoyant.io>