tls: Avoid holding a large buffer across await
The TLS detection code holds a ~512B buffer when peeking on a socket to
parse an SNI value from a client connection. This is done to avoid a
heap allocation--this optimization is probably premature.

Per a conversation with @carllerche, holding data across an `await` is a
painful anti-pattern that can have severe performance impacts, since the
generated `Future` ends up holding all of this data (on the stack).
Instead, our default behavior should be to use a heap-allocated data
structure for any data that crosses an async `await`.

This change eliminates the stack-allocated buffer in our TLS server in
favor of a heap-allocated buffer.

Signed-off-by: Oliver Gould <ver@buoyant.io>