Fix inbound server label ambiguity (#1947)
As the proxy discovers policies, it refers to the source of the policy
configuration. This comes in two flavors: "default" (synthetic)
resources and resources that actually exist in the cluster.

When the proxy reads routes and authorization, the control plane uses
protobuf types to differentiate these states; but server resources are
described by a generic label map. When the proxy reads this label map,
it doesn't properly detect when default/synthetic resources are used and
it _always_ indicates that the resource is real.

This can lead to duplicate metrics when, for instance, the proxy
initially uses a default server and then obtains a default configuration
by discovery.

This change fixes this situation in two ways:

1. We now properly detect and differentiate default server
   configurations.
2. We update the hash and comparisons functions for the `Meta` type so
   that an improperly construct resource matches a default (i.e. when
   inserted into a label map).

Fixes linkerd/linkerd2#9502

Signed-off-by: Oliver Gould <ver@buoyant.io>