[WIP] allow routes to modify request targets
This is an exploratory change to allow the profile router to modify the
target based on a request.

The router is changed so that it no longer keeps a map of routes to
services, and instead uses the route to determine a target that's passed
to the `NewService`, oneshotting the resulting `Service`. In order to
retain the current behavior where services built for routes are reused,
we would layer a `Cache` layer on the `NewService` that's passed in,
similarly to what we do for the `Router` service. Alternatively, we
could hard-code the cache into the profile router, if we always want the
caching behavior.

A new `RequestTarget` trait is added to allow constructing a target
based on a route and a request, and the initial target. This may not be
the nicest abstraction, and we can change this in the future, but this
is a sketch.

This doesn't quite compile yet, but I'm opening a draft PR for now to
show the approach.