move client policy discovery out of HTTP logical stack
This branch refactors the prototype client policy code to move the
discovery of client policies to still occur after profile discovery, but
before the HTTP logical stack. This means that SO_ORIGINAL_DST addresses
need not be stored in the `Logical` target type any longer, which I
think is significantly nicer, as we no longer have a `Logical` field
that's ignored by hashing and equality.

Instead, the client policy discovery now occurs in the `switch_logical`
stack. This is also not my _favorite_ for it, because now that stack is
responsible for more than just switching based on whether or not a
logical destination exists...but maybe the solution is just to rename
that stack.

However, with the current structure of the outbound proxy, if we want
`push_switch_logical` to push a stack that outptus a `Logical` target,
client policy discovery has to occur there, because we would want to
perform client policy lookups prior to constructing the `Logical` target
type, while we still have an `OrigDstAddr`. Alternatively, we could have
that stack output a different target type, and push a client policy
discovery stack that consumes that target type and constructs a
`Logical` target, but this felt like the simplest approach that didn't
require propagating the `OrigDstAddr` via the `Logical` target (like we
were doing previously).

Depends on #2029